fix: fixing use client bugs
